# EcommerceAPI.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "EcommerceAPI",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/register": {
      "post": {
        "summary": "register user",
        "tags": [
          "User"
        ],
        "operationId": "registeruser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/registeruserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "login user",
        "tags": [
          "User"
        ],
        "operationId": "loginuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/loginuserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/getAllUser": {
      "get": {
        "summary": "get user",
        "tags": [
          "User"
        ],
        "operationId": "getuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/4": {
      "get": {
        "summary": "get user id",
        "tags": [
          "User"
        ],
        "operationId": "getuserid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/1": {
      "put": {
        "summary": "update user",
        "tags": [
          "User"
        ],
        "operationId": "updateuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateuserrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "detele user",
        "tags": [
          "User"
        ],
        "operationId": "deteleuser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/verifyTokenEmail/eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Mywicm9sZSI6MSwiaWF0IjoxNjI2Mjk3NTUwLCJleHAiOjE2MjYzODM5NTB9.OdEPj_ki4m4t1IYyECfNEY7jEEmCtBtKDIqyrxI4Ni8": {
      "post": {
        "summary": "verify token email",
        "tags": [
          "User"
        ],
        "operationId": "verifytokenemail",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/verifytokenemailrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/flashSale/createFlashSale": {
      "post": {
        "summary": "create flash sale",
        "tags": [
          "flash sale"
        ],
        "operationId": "createflashsale",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createflashsalerequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/flashSale/getFlashSale": {
      "get": {
        "summary": "get flash Sale",
        "tags": [
          "flash sale"
        ],
        "operationId": "getflashSale",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/flashSale/getFlashSale/1": {
      "get": {
        "summary": "get flash Sale ID",
        "tags": [
          "flash sale"
        ],
        "operationId": "getflashSaleID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/flashSale/updateFlashSale/1": {
      "put": {
        "summary": "update flash sale",
        "tags": [
          "flash sale"
        ],
        "operationId": "updateflashsale",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateflashsalerequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/flashSale/notification": {
      "post": {
        "summary": "notification",
        "tags": [
          "flash sale"
        ],
        "operationId": "notification",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "registeruserrequest": {
      "title": "registeruserrequest",
      "example": {
        "username": "kiendo123",
        "email": "kiendo123@gmail.com",
        "password": "12345",
        "phonenumber": 32999995,
        "address": "Hanoi"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phonenumber": {
          "type": "integer",
          "format": "int32"
        },
        "address": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "email",
        "password",
        "phonenumber",
        "address"
      ]
    },
    "loginuserrequest": {
      "title": "loginuserrequest",
      "example": {
        "email": "kiendo123@gmail.com",
        "password": "12345"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "updateuserrequest": {
      "title": "updateuserrequest",
      "example": {
        "username": "Hoa"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        }
      },
      "required": [
        "username"
      ]
    },
    "verifytokenemailrequest": {
      "title": "verifytokenemailrequest",
      "example": {
        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Mywicm9sZSI6MSwiaWF0IjoxNjI2Mjk3NTUwLCJleHAiOjE2MjYzODM5NTB9.OdEPj_ki4m4t1IYyECfNEY7jEEmCtBtKDIqyrxI4Ni8"
      },
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "required": [
        "token"
      ]
    },
...
    "createflashsalerequest": {
      "title": "createflashsalerequest",
      "example": {
        "flashSaleName": "FS8-3",
        "startTime": "2021-07-15T03:30:00",
        "stopTime": "2021-07-15T04:55:00",
        "discount": 10,
        "amount": 12,
        "idItems": [
          1,
          2,
          3
        ]
      },
      "type": "object",
      "properties": {
        "flashSaleName": {
          "type": "string"
        },
        "startTime": {
          "type": "string"
        },
        "stopTime": {
          "type": "string"
        },
        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "idItems": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "required": [
        "flashSaleName",
        "startTime",
        "stopTime",
        "discount",
        "amount",
        "idItems"
      ]
    },
    "updateflashsalerequest": {
      "title": "updateflashsalerequest",
      "example": {
        "flashSaleName": "Fs20-11",
        "startTime": "2021-07-15T03:30:00",
        "stopTime": "2021-07-15T04:55:00",
        "discount": 10,
        "amount": 12,
        "idItems": [
          1,
          2,
          3
        ]
      },
      "type": "object",
      "properties": {
        "flashSaleName": {
          "type": "string"
        },
        "startTime": {
          "type": "string"
        },
        "stopTime": {
          "type": "string"
        },
        "discount": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "idItems": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "required": [
        "flashSaleName",
        "startTime",
        "stopTime",
        "discount",
        "amount",
        "idItems"
      ]
    },
...
  },
...
}